# Rewrite SplitLocale to not allocate