--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -22,17 +22,17 @@ jobs:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
         go:
-          - ~1.17
           - ~1.18
           - ~1.19
           - ~1.20
           - ~1.21
+          - ~1.22
     runs-on: ${{ matrix.os }}
     steps:
       - name: Checkout the code
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Set up Go
-        uses: actions/setup-go@v4
+        uses: actions/setup-go@v5
         with:
           go-version: ${{ matrix.go }}
       - name: Cache the Go modules