@Jacalz 2024-02-22
Looks like the workflows med ro be updated to run on Go 1.19.