@Jacalz 2024-02-23
I'm happy to help. You might get more contributions for the team given that we are planning on using your package in Fyne going forward :)